requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
pillow>=10.0.0
openai>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
pytest>=7.4.0
